refactor: remove legacy sidebar components and streamline imports
- Deleted the AppSidebar, NavMain, NavProjects, NavSecondary, NavUser, and Navigation components to clean up the codebase and eliminate unused files.
- Updated import paths in Layout to reflect the removal of these components, ensuring consistency and clarity in the code structure.
- This cleanup enhances maintainability and paves the way for future improvements in the navigation system.
