+++ src/components/AppSidebar.tsx
@@ -1,0 +1,109 @@
+import * as React from 'react';
+import {
+  Command,
+  LifeBuoy,
+  Send,
+  Home,
+  FolderKanban,
+  CheckSquare,
+  Users,
+  Settings,
+} from 'lucide-react';
+
+import { NavMain } from '@/components/NavMain';
+
+import { NavUser } from '@/components/NavUser';
+import {
+  Sidebar,
+  SidebarContent,
+  SidebarFooter,
+  SidebarHeader,
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
+} from '@/components/ui/sidebar';
+import { NavSecondary } from './NavSecondary';
+
+const data = {
+  user: {
+    name: 'John Doe',
+    email: 'john@example.com',
+    avatar: '/avatars/user.jpg',
+  },
+  navMain: [
+    {
+      title: 'Dashboard',
+      url: '/',
+      icon: Home,
+      isActive: true,
+    },
+    {
+      title: 'Projects',
+      url: '/projects',
+      icon: FolderKanban,
+    },
+    {
+      title: 'Tasks',
+      url: '/tasks',
+      icon: CheckSquare,
+    },
+    {
+      title: 'Team',
+      url: '/team',
+      icon: Users,
+    },
+    {
+      title: 'Settings',
+      url: '/settings',
+      icon: Settings,
+    },
+  ],
+  navSecondary: [
+    {
+      title: 'Support',
+      url: '#',
+      icon: LifeBuoy,
+    },
+    {
+      title: 'Feedback',
+      url: '#',
+      icon: Send,
+    },
+  ],
+};
+
+export const AppSidebar = ({
+  ...props
+}: React.ComponentProps<typeof Sidebar>) => {
+  return (
+    <Sidebar
+      className="top-(--header-height) h-[calc(100svh-var(--header-height))]!"
+      {...props}
+    >
+      <SidebarHeader>
+        <SidebarMenu>
+          <SidebarMenuItem>
+            <SidebarMenuButton size="lg" asChild>
+              <a href="#">
+                <div className="bg-sidebar-primary text-sidebar-primary-foreground flex aspect-square size-8 items-center justify-center rounded-lg">
+                  <Command className="size-4" />
+                </div>
+                <div className="grid flex-1 text-left text-sm leading-tight">
+                  <span className="truncate font-medium">Project Manager</span>
+                  <span className="truncate text-xs">Enterprise</span>
+                </div>
+              </a>
+            </SidebarMenuButton>
+          </SidebarMenuItem>
+        </SidebarMenu>
+      </SidebarHeader>
+      <SidebarContent>
+        <NavMain items={data.navMain} />
+        <NavSecondary className="mt-auto" />
+      </SidebarContent>
+      <SidebarFooter>
+        <NavUser user={data.user} />
+      </SidebarFooter>
+    </Sidebar>
+  );
+};
