+++ src/components/NavSecondary.tsx
@@ -1,0 +1,33 @@
+import * as React from 'react';
+import { LanguageSwitcher } from '@/components/LanguageSwitcher';
+import { ThemeToggle } from '@/components/ThemeToggle';
+
+import {
+  SidebarGroup,
+  SidebarGroupContent,
+  SidebarMenu,
+  SidebarMenuItem,
+} from '@/components/ui/sidebar';
+
+export const NavSecondary = ({
+  ...props
+}: React.ComponentPropsWithoutRef<typeof SidebarGroup>) => {
+  return (
+    <SidebarGroup {...props}>
+      <SidebarGroupContent>
+        <SidebarMenu>
+          <SidebarMenuItem>
+            <div className="flex items-center justify-center p-2">
+              <LanguageSwitcher />
+            </div>
+          </SidebarMenuItem>
+          <SidebarMenuItem>
+            <div className="flex items-center justify-center p-2">
+              <ThemeToggle />
+            </div>
+          </SidebarMenuItem>
+        </SidebarMenu>
+      </SidebarGroupContent>
+    </SidebarGroup>
+  );
+};
