+++ src/components/SiteHeader.tsx
@@ -1,0 +1,77 @@
+import { SidebarIcon, Home } from 'lucide-react';
+import { Link, useLocation } from 'react-router-dom';
+import { useTranslation } from 'react-i18next';
+
+import { SearchForm } from '@/components/search-form';
+import {
+  Breadcrumb,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  BreadcrumbList,
+  BreadcrumbPage,
+  BreadcrumbSeparator,
+} from '@/components/ui/breadcrumb';
+import { Button } from '@/components/ui/button';
+import { Separator } from '@/components/ui/separator';
+import { useSidebar } from '@/components/ui/sidebar';
+
+export const SiteHeader = () => {
+  const { toggleSidebar } = useSidebar();
+  const { t } = useTranslation();
+  const location = useLocation();
+
+  const pathSegments = location.pathname
+    .split('/')
+    .filter(Boolean)
+    .map((segment, index, array) => ({
+      name: segment,
+      path: '/' + array.slice(0, index + 1).join('/'),
+      isLast: index === array.length - 1,
+    }));
+
+  return (
+    <header className="bg-background sticky top-0 z-50 flex w-full items-center border-b">
+      <div className="flex h-(--header-height) w-full items-center gap-2 px-4">
+        <Button
+          className="h-8 w-8"
+          variant="ghost"
+          size="icon"
+          onClick={toggleSidebar}
+        >
+          <SidebarIcon />
+        </Button>
+        <Separator orientation="vertical" className="mr-2 h-4" />
+        <Breadcrumb className="hidden sm:block">
+          <BreadcrumbList>
+            <BreadcrumbItem>
+              <BreadcrumbLink asChild>
+                <Link to="/">
+                  <Home className="h-4 w-4" />
+                </Link>
+              </BreadcrumbLink>
+            </BreadcrumbItem>
+            {pathSegments.map(segment => (
+              <div key={segment.path} className="flex items-center space-x-2">
+                <BreadcrumbSeparator />
+                <BreadcrumbItem>
+                  {segment.isLast ? (
+                    <BreadcrumbPage className="capitalize">
+                      {t(`navigation.${segment.name}`)}
+                    </BreadcrumbPage>
+                  ) : (
+                    <BreadcrumbLink asChild>
+                      <Link to={segment.path} className="capitalize">
+                        {t(`navigation.${segment.name}`)}
+                      </Link>
+                    </BreadcrumbLink>
+                  )}
+                </BreadcrumbItem>
+              </div>
+            ))}
+          </BreadcrumbList>
+        </Breadcrumb>
+        <SearchForm className="w-full sm:ml-auto sm:w-auto" />
+      </div>
+    </header>
+  );
+};
